+++ src/Legendre.cpp
@@ -1,6 +1,6 @@
 #include "../include/Legendre.h"
 
-void Legendre(Matrix& pnm,Matrix& dpnm, double n, double m, double fi) {
+void Legendre(Matrix& pnm,Matrix& dpnm, int n, int m, double fi) {
     pnm = Matrix(n+1,m+1);
     dpnm = Matrix(n+1,m+1);
 
@@ -28,7 +28,7 @@ void Legendre(Matrix& pnm,Matrix& dpnm, double n, double m, double fi) {
     int k=2;
     while(1){
         for(int i=k; i <= n; i++)
-            pnm(i+1,j+1)=sqrt((2.*i+1.)/((i-j)*(i+j)))*((sqrt(2.*i-1.)*sin(fi)*pnm(i,j+1.))
+            pnm(i+1,j+1)=sqrt((2.*i+1.)/((i-j)*(i+j)))*((sqrt(2.*i-1.)*sin(fi)*pnm(i,j+1))
                 -(sqrt(((i+j-1.)*(i-j-1.))/(2.*i-3.))*pnm(i-1,j+1)));
 
         j = j+1;
